requirements.txt
dj-database-url==2.2.0
Django==5.2.4
djangorestframework==3.15.2
drf-yasg==1.21.8
Pillow==11.0.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10